requirements.txt
pandas
openpyxl
flask
flask-bcrypt
PyJWT
waitress
python-dotenv
pytest
